Consumer/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
google-cloud-language==1.1.1
PyMySQL==0.9.3
Flask-SQLAlchemy==2.3.2
python-dateutil==2.8.0
